Fix sign-compare warning in parseCommandString
  Cast *count to size_t in comparison with capacity to resolve
  -Wsign-compare error after changing capacity from int to size_t.

+++ bpv7/metadata/metadata.c
@@ -681,7 +681,7 @@ char **parseCommandString(const char *inputString, int *count)
 	{
 		stripLeadingWhiteSpace(token);
 
-		if (*count >= capacity)
+		if ((size_t)*count >= capacity)
 		{
 			capacity *= 2;
 			result = realloc(result, capacity * sizeof(char *));
